feat: read json and create user

+++ src/storage/model/user.zig
@@ -2,7 +2,7 @@ pub const User = struct {
     const Self = @This();
     name: []const u8,
     email: []const u8,
-    age: u8,
+    age: u16,
 
     pub fn create_dummy_user() User {
         return User{
